Recover sky visibility return type

+++ include/main/sky.h
@@ -99,7 +99,7 @@ void skyGetCurrentAmbientAndLightColors(u8* ambientRed, u8* ambientGreen, u8* am
                                         u8* lightBlue);
 void* fn_8008912C(void);
 void skyBuildSunModelMatrix(f32 mtx[3][4]);
-int skyFn_8008919c(int slot);
+u8 skyFn_8008919c(int slot);
 void getTimeOfDay(f32* time);
 void renderSky(int a, int b, int c, int d, int visible);
 void getAmbientColor(int slot, u8* red, u8* green, u8* blue);
